requirements.txt
Django==3.2
djangorestframework==3.14.0
django-phonenumber-field==7.1.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0
django-allauth==0.54.0
django-environ==0.10.0
psycopg2-binary==2.9.6
Pillow==9.5.0
drf_spectacular==0.26.2
drf-extra-fields==3.4.1
drf-psq~=1.1.0
django-cleanup==7.0.0
celery==5.2.7
redis==4.5.4
django_celery_beat==2.5.0
python-dateutil~=2.8.2
django_filter==23.2
pytest_django~=4.5.2
Faker==18.6.0
gunicorn~=20.1.0
drf_spectacular_sidecar==2023.5.1
phonenumbers==8.13.11
pytest~=7.3.1
geopy~=2.3.0